* fixed: changed std::stoi to std::stoll

+++ main.cpp
@@ -64,11 +64,11 @@ int main(const int argc, const char *argv[]) {
         serial_alg_string   = parse_config_to_serial(config, n_layers);
         parallel_alg_string = parse_config_to_parallel(config, n_layers);
 
-        n_threads                = std::stoi(clp.get("--threads"));
+        n_threads                = std::stoll(clp.get("--threads"));
         parallel_strategy_string = clp.get("--strategy");
 
         if (clp.is_set("--seed")) {
-            seed = std::stoi(clp.get("--seed"));
+            seed = std::stoll(clp.get("--seed"));
         }
     }
